requirements.txt
aiofiles~=23.2.1
bcrypt
fastapi~=0.111.0
passlib
pydantic~=2.7.2
pymongo~=4.7.2
python-decouple~=3.8
python-jose
python-multipart